centralize all types in types-only file, and simplify `index.d.ts` consumption
